feat: Enhance MQTT and OTA functionality
- Updated HiveMqCert.h to include additional trust anchors for improved security.
- Modified LedConfigStorage.h to align default LED on/off times with new requirements.
- Enhanced MqttBridge.cpp with additional error handling and support for OTA Slave updates.
- Introduced new SlaveOtaProxy class for managing OTA updates for Slave devices.
- Improved resilience mechanisms in SlaveResilience.h to handle idle states and heap monitoring.
- Added stuck relay detection in VentilationZone to ensure reliable operation.
- Implemented redirect handling in OtaUpdater to manage HTTP redirects during OTA.
- Updated TimeSync.h to ensure proper uptime reporting.
- Refined logging and error handling across various components for better diagnostics.

+++ ESP32/TimeSync.h
@@ -87,7 +87,7 @@ public:
             return true;
         }
         // Fallback: "uptime:NNNNs"
-        snprintf(buf, bufLen, "uptime:%us", epochSec);
+        snprintf(buf, bufLen, "uptime:%us", (unsigned int)epochSec);
         return false;
     }
